+++ App/src/model/state/LocationState.ts
@@ -3,6 +3,8 @@ import { LocationData } from 'expo-location';
 
 export type LocationState = {
   track: boolean,
+
+  timestamp?: number,
   location?: LocationData,
   address?: Location.Address,
 };