+++ App/src/screens/More/Nearby/styles.tsx
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    emptyContainer: {
+        flex: 1,
+        // height: Dimensions.get('window').height,
+        alignItems: 'center',
+        justifyContent: 'center',
+        padding: 16,
+    },
+    emptyText: {
+        paddingVertical: 16,
+        fontSize: 20,
+        textAlign: 'center',
+    },
+});